Modulo-06-Funcionalidades/31-Codigo-Barra-QR/requirements.txt
colorama==0.4.5
Pillow==9.2.0
python-barcode==0.14.0
qrcode==7.3.1
